Fix slow page loading
The slowness was caused by the app trying to render all 59 thousand gene
names in the dropdown. This was previously limited to 50, as a user will
(hopefully!) not need to manually browse all 59k genes, instead they
will search for the genes they are interested in, which hopefully should
be returned in the 50 that are displayed.
This limit was removed in a previous commit