feat: data upload (#49)
* Python script testing (specifically, working out why it's causing WSL to crash)

* Edits data_upload script to include chunksize (fixing out-of-memory crash in WSL) and replace data in tables rather than append - first steps in making script work for updates which include removing datasets

* adds commandline input for paths to raw data, metadata, and database location

* Removed commented out lines

* House keeping

* Look for raw files in supplied dir

`merge_all_datasets` was previously loading in a 'master' file, which was a TSV
file that held a list of filenames of raw files to load into the DB.
This is an additional layer of metadata that is unnecessary, instead the
script now looks for any file in the specified folder and attempts to
insert it into the DB.

Additional changes to `insert_into_db` are:

  - Reducing `chunksize` down to 5,000, as I was still getting
    out-of-memory crashes occasionally
  - Changing the `if_exists` behaviour to `fail`. The script should only
  ever be run on an empty database, so if any of the tables do exist
  already, this is a failure condition

* Removed fix for the Uro-D and Uro-P0 samples

This is instead handled in the metadata

* Removed lines that generate unused variables

* Don't remove hardcoded samples from gene data

Instead, these samples should be removed from the metadata (which they
have already been), to preserve the metadata as the ground source of
truth. The script now only keeps gene expression measurements for
samples with a corresponding SampleId in the metadata (this will also
stop any errors of violating a FK constraint)

* WIP: Use common function for dimension table prep

NB: this isn't entirely working as expected, as it's creating larger DBs
than before the change. This needs to be investigated

* Fix filtering expression data to metadata SampleIds

Previously the filter was achieved by an inner join between the 2
tables. However, I hadn't realised that the `pandas.merge` function adds
the joining column from the right hand table to the result, which was
unintentional. The filtering is now done by `isin`.

* Created launch shell script

This shell script is what will be run when Puppet detects a change in
the raw data / metadata

* Fix: foreign key relationships are obeyed

* Positively select metadata cols rather than negatively

We want to ensure that we always have the required metadata columns,
this is best achieved by explicitly selecting those columns. However,
previously this was done negatively (i.e. by removing unneeded columns).

---------

Co-authored-by: Finn <finn.mclellan@york.ac.uk>