Chunked melt and DB insert operation to save memory usage.
The wide table (genes on rows, samples in columns) is chunked in
both dimensions. Each chunk is pivoted to long to match the DB
table format, and then the long chunk is inserted into the DB.

This saves a considerable amount of memory over making a single long
table (chunked or not) and inserting that in 1 go (chunked or not).