feat: working DB build script (#50)
* Variables passed as arguments

* Remove requirement that metadata is within data folder

It might be useful to separate the metadata file from the raw data
folder to keep separation of concerns, and also because it isn't obvious
unless you look in the script that this is the case.

* Updates python to python3

* Chunk melt operation

The melt call was resulting in high memory usage, going OOM on the VM.
Chunking the melt will reduce memory usage, hopefully allowing this
script to run successfully.

* Chunked melt and DB insert operation to save memory usage.

The wide table (genes on rows, samples in columns) is chunked in
both dimensions. Each chunk is pivoted to long to match the DB
table format, and then the long chunk is inserted into the DB.

This saves a considerable amount of memory over making a single long
table (chunked or not) and inserting that in 1 go (chunked or not).

* Only print success message when asked

---------

Co-authored-by: Philip Harrison <philip.harrison@york.ac.uk>